backend/requirements.txt
# Vérifié via safety-mcp (latest_secure / recommandé) — mars 2026
pytest==9.0.2
OpenHosta==4.0.2
python-dotenv==1.2.2
fastapi==0.135.1
uvicorn[standard]==0.42.0
httpx==0.28.1
PyYAML==6.0.3
mcp==1.26.0
